pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "schemas*", "services*", "repositories*", "utils*", "tests*"]

[project]
name = "community-recipe-book"
version = "0.1.0"
description = "Community Recipe Book - A full-stack web application for sharing recipes"
requires-python = ">=3.8"

authors = [
    { name = "Denyss", email = "carcamo.denyss@gmail.com" },
    { name = "Damaris", email = "x@gmail.com" },
    { name = "Eduardo", email = "eduardo.abarca.c@usach.cl" }
]

dependencies = [
    "annotated-types==0.7.0",
    "backoff==2.2.1",
    "blinker==1.9.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "email-validator==2.3.0",
    "Flask==3.1.2",
    "flask-cors==5.0.0",
    "flasgger==0.9.7.1",
    "greenlet==3.2.4",
    "idna==3.10",
    "importlib_metadata==8.7.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "mysql-connector-python==9.4.0",
    "pydantic==2.12.0",
    "pydantic_core==2.41.1",
    "PyJWT==2.10.1",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
    "requests==2.32.4",
    "SQLAlchemy==2.0.44",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "Werkzeug==3.1.3",
    "zipp==3.23.0"
]

[project.optional-dependencies]
dev = [
    "pylint",
    "pre-commit",
    "pytest",
    "pytest-flask",
    "pytest-cov"
]
